Updated di version. Negative delta values are functional in the controller.

+++ library/source/input/input_system.cpp
@@ -32,14 +32,14 @@ void input_system::update (frame_timer::duration delta, scene* scene)
     auto transform  = entity->component<mak::transform> ();
     if (std::find(controller_cache_.begin(), controller_cache_.end(), controller) == controller_cache_.end())
     {
-      if(controller->on_key_press    ) on_key_press       .connect([=] (di::key                    key   ) { controller->on_key_press    (transform, key   ); });
-      if(controller->on_key_down     ) on_key_down        .connect([=] (di::key                    key   ) { controller->on_key_down     (transform, key   ); });
-      if(controller->on_key_release  ) on_key_release     .connect([=] (di::key                    key   ) { controller->on_key_release  (transform, key   ); });
-      if(controller->on_mouse_press  ) on_mouse_press     .connect([=] (std::size_t                button) { controller->on_mouse_press  (transform, button); });
-      if(controller->on_mouse_down   ) on_mouse_down      .connect([=] (std::size_t                button) { controller->on_mouse_down   (transform, button); });
-      if(controller->on_mouse_release) on_mouse_release   .connect([=] (std::size_t                button) { controller->on_mouse_release(transform, button); });
-      if(controller->on_mouse_move   ) on_mouse_move_delta.connect([=] (std::array<std::size_t, 2> value ) { controller->on_mouse_move   (transform, value ); });
-      if(controller->on_mouse_wheel  ) on_mouse_wheel     .connect([=] (std::array<std::size_t, 2> value ) { controller->on_mouse_wheel  (transform, value ); });
+      if(controller->on_key_press    ) on_key_press       .connect([=] (di::key                     key   ) { controller->on_key_press    (transform, key   ); });
+      if(controller->on_key_down     ) on_key_down        .connect([=] (di::key                     key   ) { controller->on_key_down     (transform, key   ); });
+      if(controller->on_key_release  ) on_key_release     .connect([=] (di::key                     key   ) { controller->on_key_release  (transform, key   ); });
+      if(controller->on_mouse_press  ) on_mouse_press     .connect([=] (std::size_t                 button) { controller->on_mouse_press  (transform, button); });
+      if(controller->on_mouse_down   ) on_mouse_down      .connect([=] (std::size_t                 button) { controller->on_mouse_down   (transform, button); });
+      if(controller->on_mouse_release) on_mouse_release   .connect([=] (std::size_t                 button) { controller->on_mouse_release(transform, button); });
+      if(controller->on_mouse_move   ) on_mouse_move_delta.connect([=] (std::array<std::int32_t, 2> value ) { controller->on_mouse_move   (transform, value ); });
+      if(controller->on_mouse_wheel  ) on_mouse_wheel     .connect([=] (std::array<std::int32_t, 2> value ) { controller->on_mouse_wheel  (transform, value ); });
       controller_cache_.push_back(controller);
     }
   }
